Reports - use right path for outputSummary job templates
